# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gestão de despesas em grupo.",
        "title": "TripSplit API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Autentica um usuário e retorna um token JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login de usuário",
                "parameters": [
                    {
                        "description": "Dados de login",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Cria uma nova conta de usuário com nome, e-mail e senha.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Cadastrar novo usuário",
                "parameters": [
                    {
                        "description": "Dados do usuário",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/expenses": {
            "post": {
                "description": "Registra uma nova despesa associada a um grupo, com detalhes de quem pagou e como a despesa deve ser dividida entre os membros.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Criar nova despesa",
                "parameters": [
                    {
                        "description": "Dados da despesa",
                        "name": "expense",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http.CreateExpenseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_pairbytes-dev_tripsplit-monorepo_internal_core_domain.Expense"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/groups": {
            "post": {
                "description": "Cria um novo grupo de despesas, associando-o ao usuário autenticado como proprietário.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Criar novo grupo",
                "parameters": [
                    {
                        "description": "Dados do grupo",
                        "name": "group",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_http.createGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_pairbytes-dev_tripsplit-monorepo_internal_core_domain.Group"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_pairbytes-dev_tripsplit-monorepo_internal_core_domain.Expense": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "method": {
                    "description": "equal, percentage, fixed",
                    "type": "string"
                },
                "payer_id": {
                    "type": "string"
                },
                "splits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_pairbytes-dev_tripsplit-monorepo_internal_core_domain.ExpenseSplit"
                    }
                }
            }
        },
        "github_com_pairbytes-dev_tripsplit-monorepo_internal_core_domain.ExpenseSplit": {
            "type": "object",
            "properties": {
                "amount_owed": {
                    "type": "number"
                },
                "expense_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_pairbytes-dev_tripsplit-monorepo_internal_core_domain.Group": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_pairbytes-dev_tripsplit-monorepo_internal_core_domain.Expense"
                    }
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_pairbytes-dev_tripsplit-monorepo_internal_core_domain.GroupMember"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_pairbytes-dev_tripsplit-monorepo_internal_core_domain.GroupMember": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "joined_at": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_http.CreateExpenseRequest": {
            "type": "object",
            "required": [
                "amount",
                "group_id",
                "payer_id",
                "splits",
                "title"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 100.5
                },
                "group_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "payer_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "splits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_http.SplitRequest"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "Almoço"
                }
            }
        },
        "internal_http.SplitRequest": {
            "type": "object",
            "required": [
                "amount",
                "user_id"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 50.25
                },
                "user_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
        "internal_http.createGroupRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Grupo para organizar uma viagem"
                },
                "title": {
                    "type": "string",
                    "example": "Grupo de Viagem"
                }
            }
        },
        "internal_http.loginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "teste@teste.com"
                },
                "password": {
                    "type": "string",
                    "example": "senha12345"
                }
            }
        },
        "internal_http.registerRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "teste@teste.com"
                },
                "name": {
                    "type": "string",
                    "example": "Teste teste"
                },
                "password": {
                    "type": "string",
                    "example": "senha12345"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}